Use a double-rename procedure for saving (i.e., create ".old" and ".new")

+++ avogadro/src/mainwindow.cpp
@@ -1134,8 +1134,15 @@ namespace Avogadro
     } // end saving settings for CML files
 
     if ( conv.Write( &obmol, &ofs ) ) {
-      file.remove(); // remove the old file: WARNING -- would much prefer to just rename, but Qt won't let you
-      newFile.rename(fileName);
+      bool success;
+      success = file.rename(fileName + ".old");
+      if (success)
+        success = newFile.rename(fileName);
+      else
+        statusBar()->showMessage( tr("Saving molecular file failed."), 5000 );
+      
+      if (success) // renaming worked
+        file.remove(); // remove the old file: WARNING -- would much prefer to just rename, but Qt won't let you
       statusBar()->showMessage( tr("Save succeeded."), 5000 );
       setWindowModified( false );
     }
